Create Brach.py
This project aims to provide Python, Verilog, C++, or another type of code to perform forward chaining in the design. This model will demonstrate how to improve the design's efficacy by applying the Forward Chaining process. This method makes the most of the hardware architecture's capability.
This model will showcase how using the Forward Chaining methodology could enhance the effectiveness of the design. This technique utilizes the potential of the hardware Architecture to its maximum to give optimum speed and better time management as an output it also demonstrates the working of 2 bit branch predictor.